Switch back to Oscar and update loading logic
The debug kit returns to using the standard `&Oscar.run` rather than the three
paragraphs of logic I had previously. This is made possible by the deep merging
behavior added to vagrant-config_builder which was released in 0.7.0.

The version-controlled config data, in `config/debugging-kit` has been
re-located to `data/puppet-debugging-kit`. The logic behind this is that the
`data/<name>` directory is accessible through `Gem.datadir` --- so this
location should be relatively future-proof if the debugging kit is ever
distributed as a Vagrant plugin.