Add version checks for required plugins
The debugging kit may mis-behave if certian plugins do not meet minimum version
requirements. This mostly arises when a user has an existing global Vagrant
installation from vagrantup.com. Starting with Vagrant 1.5.x, it is possible to
specify minimum requirements with `vagrant plugin install`. However 1.5.x is
new enough that we can't rely on everyone using it.

So, this patch adds a utility module that contains version number sanity
checks.