New Filter: debug-kit
This Filter processes any VM definition which has a `debug-kit` key:

    ---
    vms:

      - name: pe-323-master
        debug-kit: true
        ...

When activated, the filter parses the `name` and uses the information
contained therein to configure a `pe_bootstrap` provisioner for the VM.
The filter sets default values for the VM `hostname` and the `role`, `version`
and `master` configuration values of the `pe_boostrap` provisioner. Any
existing values for these attributes that have been set directly in the VM
definition or inherited from a role will be left intact by the filter.