Add memory tuning to PE 3.7.2 and newer
This commit adds one new role: `pe-memory-tuning`

Applying this role prior to installing a PE master will add defaults that allow
the PE infrastructure to operate on a VM with only 1 GB of RAM allocated.

PE 3.7.2 or newer is required due to lack of support for tuning JRuby in 3.7.0
and 3.7.1.