Update base role with better hosts options
This patch sets `sync_hosts` to true which causes any update made by the
`hosts` provisioner to propagate to all running VMs. This change requires
vagrant-hosts version 2.6.0.

The `add_localhost_hostnames` option is also set to `false`. This disables the
extraneous `127.0.1.1` entry which was misguided and will be either altered or
removed in an upcoming major version of vagrant-hosts.