Handle removal of Vagrant boxes from Atlas
Hosting of Vagrant boxes on atlas.hashicorp.com has moved to vagrantcloud.com
and this functionality of atlas has been decommissioned, apparently some time
last week. This means that VM creation fails when a box has to be downloaded.

Vagrant migrated to using vagrantcloud.com as the default server URL in version
1.9.6, but Oscar is currently stuck on 1.8.7 due to testing issues. This commit
adds a conditional to the Debug Kit plugin checks run during startup that
patches in an updated URL if the Vagrant version is too old.

Fixes Sharpie/puppet-debugging-kit#19.