Add 128 MB to the Puppet Server memory allotment
PE 2018.1 switches to using JRuby 9.x by default instead of the 1.7.x version
that has been used for many years. JRuby 9.x requires a bit more RAM to operate
so this commit adds 128 MB to the memory limit for the Puppet Server service.
This was the amount required for back-to-back-to-back compilations of a
monolithic master catalog without an OOM event ocurring.