(maint) Reduce Jruby 9k code cache size
Prior to this commit PE 2018.1+ would add a 512MB code cache to the
Jruby 9k instance. As this could cause OS OOMs, it should be reduced
to a smaller size. This commit reduces the code cache size to 96MB
in order to use the memory previously allocated to activemq, which is
not installed in the 2018.1 release by default.