(SUP-708) Remove providers for compile and HA examples
Prior to this commit, there were specific providers in the compile and
HA examples. This commit removes the providers and adds back in the
memory tuning for the compile example. The memory tuning was not
configured in the HA example due to the large size and that code manager
will automatically be configured and immediately overwrite it.