Update memory tuning role for 2018.1 and 2019.0
PE 2018.1 changed `puppet infrastructure configure` to run in a special
environment named "enterprise". This means that Hiera data with tuning
parameters was no longer sourced from the production environment. This
commit fixes tuning by creating an "enterprise" environment and
populating it with data. A FIXME has also been added as we should dump
this eventually and just add the tuning to pe.conf.